requirements.txt
streamlit
numpy
requests
faiss-cpu
tqdm
pypdf